requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.11.7
pydantic-settings==2.1.0
pyyaml==6.0.1
httpx==0.27.0
pytest==8.2.2
pytest-asyncio==0.23.2
xmltodict==0.14.2
